requirements.txt
# Core dependencies with flexible version ranges for better compatibility
locust>=2.20.0,<3.0.0
requests>=2.28.0,<3.0.0
PyYAML>=6.0,<7.0
pycryptodome>=3.18.0,<4.0.0

# Locust dependencies (will be installed automatically but listed for clarity)
Flask>=3.0.0,<4.0.0
gevent>=23.0.0,<25.0.0
psutil>=5.9.0,<6.0.0
